Move table functions to own file to avoid import errors.
